+++ src/one-source.js
@@ -123,7 +123,7 @@ withOneSource('transduce', {
 
 
 var withFnArgMixin = {
-  _init: function(args) {  this._fn = args[0]  },
+  _init: function(args) {  this._fn = args[0] || id  },
   _free: function() {  this._fn = null  }
 };
 
@@ -238,7 +238,7 @@ withOneSource('skipDuplicates', {
 
 withOneSource('skipWhile', {
   _init: function(args) {
-    this._fn = args[0];
+    this._fn = args[0] || id;
     this._skip = true;
   },
   _free: function() {